Update README.md with new sections and links
Replaced the "Hello World" section with a link to a blog post about publishing a NuGet package using GitHub. Added a "Documentation Ref" section with links to the blog post and NuGet API key application page. Introduced a "MinVer" section with a link to the MinVer GitHub repository and instructions for versioning and releasing in GitLab.